Refactor the code by using some helper functions.
The command line determines whether the user wants to upload
one or many files. It calls the corresponding function.
From there the function check for file existence, determine the
host and check for appropriate file size. It then upload the
file(s) and prints the link to the uploaded files.

At the moment, only imgur is working. Need to work on the 2 other
hosts.